Refactor quantity handling: update types to double, add localization for quantity dialog and save button

+++ lib/features/billing/presentation/bloc/billing_bloc.dart
@@ -114,7 +114,7 @@ class BillingBloc extends Bloc<BillingEvent, BillingState> {
               productName: cartItem.product.name,
               price: cartItem.product.price,
               cost: cartItem.product.cost,
-              quantity: cartItem.quantity.toDouble(),
+              quantity: cartItem.quantity,
             ))
         .toList();
